LC-656: FileReference refactor WIP, need to merge...
