+++ lucille-core/src/main/java/com/kmwllc/lucille/connector/storageclient/GoogleStorageClient.java
@@ -11,7 +11,6 @@ import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.Storage.BlobListOption;
 import com.google.cloud.storage.StorageOptions;
 import com.kmwllc.lucille.connector.FileConnector;
-import com.kmwllc.lucille.connector.storageclient.filereference.FileReference;
 import com.kmwllc.lucille.core.Document;
 import com.kmwllc.lucille.core.Publisher;
 import com.typesafe.config.Config;
@@ -27,6 +26,87 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class GoogleStorageClient extends BaseStorageClient {
+
+  public class GoogleFileReference extends BaseFileReference {
+
+    private final Blob blob;
+
+    public GoogleFileReference(Blob blob) {
+      super(blob.getUpdateTimeOffsetDateTime().toInstant());
+
+      this.blob = blob;
+    }
+
+    @Override
+    public String getFullPath(TraversalParams params) {
+      URI paramsURI = params.getURI();
+      return paramsURI.getScheme() + "://" + paramsURI.getAuthority() + "/" + blob.getName();
+    }
+
+    @Override
+    public boolean isCloudFileReference() {
+      return true;
+    }
+
+    @Override
+    public boolean isValidFile() {
+      return !blob.isDirectory();
+    }
+
+    @Override
+    public InputStream getContentStream(TraversalParams params) {
+      ReadChannel readChannel = blob.reader();
+      return Channels.newInputStream(readChannel);
+    }
+
+    @Override
+    public Document toDoc(TraversalParams params) {
+      String fullPath = getFullPath(params);
+      String docId = DigestUtils.md5Hex(fullPath);
+      Document doc = Document.create(params.getDocIdPrefix() + docId);
+
+      doc.setField(FileConnector.FILE_PATH, fullPath);
+
+      if (blob.getUpdateTimeOffsetDateTime() != null) {
+        doc.setField(FileConnector.MODIFIED, blob.getUpdateTimeOffsetDateTime().toInstant());
+      }
+
+      if (blob.getCreateTimeOffsetDateTime() != null) {
+        doc.setField(FileConnector.CREATED, blob.getCreateTimeOffsetDateTime().toInstant());
+      }
+
+      doc.setField(FileConnector.SIZE, blob.getSize());
+
+      if (params.shouldGetFileContent()) {
+        doc.setField(FileConnector.CONTENT, blob.getContent());
+      }
+
+      return doc;
+    }
+
+    @Override
+    public Document toDoc(InputStream in, String decompressedFullPathStr, TraversalParams params) throws IOException {
+      final String docId = DigestUtils.md5Hex(decompressedFullPathStr);
+      final Document doc = Document.create(params.getDocIdPrefix() + docId);
+
+      doc.setField(FileConnector.FILE_PATH, decompressedFullPathStr);
+
+      if (blob.getUpdateTimeOffsetDateTime() != null) {
+        doc.setField(FileConnector.MODIFIED, blob.getUpdateTimeOffsetDateTime().toInstant());
+      }
+
+      if (blob.getCreateTimeOffsetDateTime() != null) {
+        doc.setField(FileConnector.CREATED, blob.getCreateTimeOffsetDateTime().toInstant());
+      }
+
+      // unable to get decompressed file size
+      if (params.shouldGetFileContent()) {
+        doc.setField(FileConnector.CONTENT, in.readAllBytes());
+      }
+
+      return doc;
+    }
+  }
 
   private static final Logger log = LoggerFactory.getLogger(GoogleStorageClient.class);
   private Storage storage;
@@ -70,13 +150,10 @@ public class GoogleStorageClient extends BaseStorageClient {
     do {
       page.streamAll()
           .forEachOrdered(blob -> {
-            FileReference fileRef = new FileReference(blob);
-
-            if (validFile(fileRef)) {
-              String fullPathStr = getFullPath(blob, params);
-              String fileExtension = FilenameUtils.getExtension(fullPathStr);
-              tryProcessAndPublishFile(publisher, fullPathStr, fileExtension, fileRef, params);
-            }
+            FileReference fileRef = new GoogleFileReference(blob);
+            String fullPathStr = fileRef.getFullPath(params);
+            String fileExtension = FilenameUtils.getExtension(fullPathStr);
+            tryProcessAndPublishFile(publisher, fullPathStr, fileExtension, fileRef, params);
           });
       page = page.hasNextPage() ? page.getNextPage() : null;
     } while (page != null);
@@ -93,35 +170,6 @@ public class GoogleStorageClient extends BaseStorageClient {
   }
 
   @Override
-  protected Document convertFileReferenceToDoc(FileReference fileReference, TraversalParams params) {
-    Blob blob = fileReference.getBlob();
-
-    try {
-      return blobToDoc(blob, params);
-    } catch (IOException e) {
-      throw new IllegalArgumentException("Unable to convert blob '" + blob.getName() + "' to Document", e);
-    }
-  }
-
-  @Override
-  protected Document convertFileReferenceToDoc(FileReference fileReference, InputStream in, String decompressedFullPathStr, TraversalParams params) {
-    Blob blob = fileReference.getBlob();
-
-    try {
-      return blobToDoc(blob, in, decompressedFullPathStr, params);
-    } catch (IOException e) {
-      throw new IllegalArgumentException("Unable to convert blob '" + blob.getName() + "' to Document", e);
-    }
-  }
-
-  @Override
-  protected InputStream getFileReferenceContentStream(FileReference fileReference, TraversalParams params) {
-    Blob blob = fileReference.getBlob();
-    ReadChannel readChannel = blob.reader();
-    return Channels.newInputStream(readChannel);
-  }
-
-  @Override
   protected String getStartingDirectory(TraversalParams params) {
     URI pathURI = params.getURI();
     String startingDirectory = Objects.equals(pathURI.getPath(), "/") ? "" : pathURI.getPath();
@@ -134,67 +182,6 @@ public class GoogleStorageClient extends BaseStorageClient {
   @Override
   protected String getBucketOrContainerName(TraversalParams params) {
     return params.getURI().getAuthority();
-  }
-
-  @Override
-  protected boolean validFile(FileReference fileRef) {
-    Blob blob = fileRef.getBlob();
-    return !blob.isDirectory();
-  }
-
-  @Override
-  protected String getFullPath(FileReference fileRef, TraversalParams params) {
-    return getFullPath(fileRef.getBlob(), params);
-  }
-
-  private String getFullPath(Blob blob, TraversalParams params) {
-    return params.getURI().getScheme() + "://" + getBucketOrContainerName(params) + "/" + blob.getName();
-  }
-
-  private Document blobToDoc(Blob blob, TraversalParams params) throws IOException {
-    String fullPath = getFullPath(blob, params);
-    String docId = DigestUtils.md5Hex(fullPath);
-    Document doc = Document.create(params.getDocIdPrefix() + docId);
-
-    doc.setField(FileConnector.FILE_PATH, fullPath);
-
-    if (blob.getUpdateTimeOffsetDateTime() != null) {
-      doc.setField(FileConnector.MODIFIED, blob.getUpdateTimeOffsetDateTime().toInstant());
-    }
-
-    if (blob.getCreateTimeOffsetDateTime() != null) {
-      doc.setField(FileConnector.CREATED, blob.getCreateTimeOffsetDateTime().toInstant());
-    }
-
-    doc.setField(FileConnector.SIZE, blob.getSize());
-
-    if (params.shouldGetFileContent()) {
-      doc.setField(FileConnector.CONTENT, blob.getContent());
-    }
-
-    return doc;
-  }
-
-  private Document blobToDoc(Blob blob, InputStream content, String decompressedFullPathStr, TraversalParams params) throws IOException {
-    final String docId = DigestUtils.md5Hex(decompressedFullPathStr);
-    final Document doc = Document.create(params.getDocIdPrefix() + docId);
-
-    doc.setField(FileConnector.FILE_PATH, decompressedFullPathStr);
-
-    if (blob.getUpdateTimeOffsetDateTime() != null) {
-      doc.setField(FileConnector.MODIFIED, blob.getUpdateTimeOffsetDateTime().toInstant());
-    }
-
-    if (blob.getCreateTimeOffsetDateTime() != null) {
-      doc.setField(FileConnector.CREATED, blob.getCreateTimeOffsetDateTime().toInstant());
-    }
-
-    // unable to get decompressed file size
-    if (params.shouldGetFileContent()) {
-      doc.setField(FileConnector.CONTENT, content.readAllBytes());
-    }
-
-    return doc;
   }
 
   // Only for testing
