+++ lucille-core/src/main/java/com/kmwllc/lucille/connector/storageclient/LocalStorageClient.java
@@ -6,8 +6,6 @@ import static com.kmwllc.lucille.connector.FileConnector.FILE_PATH;
 import static com.kmwllc.lucille.connector.FileConnector.MODIFIED;
 import static com.kmwllc.lucille.connector.FileConnector.SIZE;
 
-import com.kmwllc.lucille.connector.storageclient.filereference.FileReference;
-import com.kmwllc.lucille.core.ConnectorException;
 import com.kmwllc.lucille.core.Document;
 import com.kmwllc.lucille.core.Publisher;
 
@@ -24,12 +22,99 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.time.Instant;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class LocalStorageClient extends BaseStorageClient {
+
+  private class LocalFileReference extends BaseFileReference {
+
+    private final Path path;
+
+    // The provided path should be absolute and normalized.
+    public LocalFileReference(Path path, Instant lastModified) {
+      super(lastModified);
+
+      this.path = path;
+    }
+
+    @Override
+    public String getFullPath(TraversalParams params) {
+      return path.toString();
+    }
+
+    @Override
+    public boolean isCloudFileReference() {
+      return false;
+    }
+
+    @Override
+    public boolean isValidFile() {
+      return true;
+    }
+
+    @Override
+    public InputStream getContentStream(TraversalParams params) {
+      try {
+        return Files.newInputStream(path);
+      } catch (IOException e) {
+        throw new IllegalArgumentException("Unable to get content stream of path '" + path + "'", e);
+      }
+    }
+
+    @Override
+    public Document toDoc(TraversalParams params) {
+      String fullPath = path.toAbsolutePath().normalize().toString();
+      String docId = DigestUtils.md5Hex(fullPath);
+      Document doc = Document.create(createDocId(docId, params));
+
+      try {
+        BasicFileAttributes attrs = Files.readAttributes(path, BasicFileAttributes.class);
+
+        doc.setField(FILE_PATH, fullPath);
+        doc.setField(SIZE, attrs.size());
+
+        if (attrs.lastModifiedTime() != null) {
+          doc.setField(MODIFIED, attrs.lastModifiedTime().toInstant());
+        }
+
+        if (attrs.creationTime() != null) {
+          doc.setField(CREATED, attrs.creationTime().toInstant());
+        }
+
+        if (params.shouldGetFileContent()) {
+          doc.setField(CONTENT, Files.readAllBytes(path));
+        }
+      } catch (IOException e) {
+        throw new IllegalArgumentException("Unable to convert path '" + path + "' to Document", e);
+      }
+      return doc;
+    }
+
+    @Override
+    public Document toDoc(InputStream in, String decompressedFullPathStr, TraversalParams params) throws IOException {
+      String docId = DigestUtils.md5Hex(decompressedFullPathStr);
+      Document doc = Document.create(createDocId(docId, params));
+
+      // get file attributes
+      BasicFileAttributes attrs = Files.readAttributes(path, BasicFileAttributes.class);
+
+      // setting fields on document
+      // remove Extension to show that we have decompressed the file and obtained its information
+      doc.setField(FILE_PATH, decompressedFullPathStr);
+      doc.setField(MODIFIED, attrs.lastModifiedTime().toInstant());
+      doc.setField(CREATED, attrs.creationTime().toInstant());
+      // unable to get decompressed file size
+      if (params.shouldGetFileContent()) {
+        doc.setField(CONTENT, in.readAllBytes());
+      }
+
+      return doc;
+    }
+  }
 
   private static final Logger log = LoggerFactory.getLogger(LocalStorageClient.class);
 
@@ -39,16 +124,13 @@ public class LocalStorageClient extends BaseStorageClient {
 
   // Config options do not matter for LocalStorageClient
   @Override
-  protected void validateOptions(Config config) {
-  }
+  protected void validateOptions(Config config) { }
 
   @Override
-  protected void initializeStorageClient() throws IOException {
-  }
+  protected void initializeStorageClient() throws IOException { }
 
   @Override
-  protected void shutdownStorageClient() throws IOException {
-  }
+  protected void shutdownStorageClient() throws IOException { }
 
   @Override
   protected void traverseStorageClient(Publisher publisher, TraversalParams params) throws Exception {
@@ -63,37 +145,6 @@ public class LocalStorageClient extends BaseStorageClient {
   }
 
   @Override
-  protected Document convertFileReferenceToDoc(FileReference fileReference, TraversalParams params) {
-    Path path = fileReference.getPath();
-    try {
-      return pathToDoc(path, params);
-    } catch (ConnectorException e) {
-      throw new IllegalArgumentException("Unable to convert path '" + path + "' to Document", e);
-    }
-  }
-
-  @Override
-  protected Document convertFileReferenceToDoc(FileReference fileReference, InputStream in, String decompressedFullPathStr,
-      TraversalParams params) {
-    Path path = fileReference.getPath();
-    try {
-      return pathToDoc(path, in, decompressedFullPathStr, params);
-    } catch (ConnectorException e) {
-      throw new IllegalArgumentException("Unable to convert path '" + path + "' to Document", e);
-    }
-  }
-
-  @Override
-  protected InputStream getFileReferenceContentStream(FileReference fileReference, TraversalParams params) {
-    Path path = fileReference.getPath();
-    try {
-      return Files.newInputStream(path);
-    } catch (IOException e) {
-      throw new IllegalArgumentException("Unable to get content stream of path '" + path + "'", e);
-    }
-  }
-
-  @Override
   protected String getStartingDirectory(TraversalParams params) {
     return params.getURI().getPath();
   }
@@ -101,69 +152,6 @@ public class LocalStorageClient extends BaseStorageClient {
   @Override
   protected String getBucketOrContainerName(TraversalParams params) {
     return params.getURI().getAuthority();
-  }
-
-  private Document pathToDoc(Path path, TraversalParams params) throws ConnectorException {
-    String fullPath = path.toAbsolutePath().normalize().toString();
-    String docId = DigestUtils.md5Hex(fullPath);
-    Document doc = Document.create(createDocId(docId, params));
-
-    try {
-      BasicFileAttributes attrs = Files.readAttributes(path, BasicFileAttributes.class);
-
-      doc.setField(FILE_PATH, fullPath);
-      doc.setField(SIZE, attrs.size());
-
-      if (attrs.lastModifiedTime() != null) {
-        doc.setField(MODIFIED, attrs.lastModifiedTime().toInstant());
-      }
-
-      if (attrs.creationTime() != null) {
-        doc.setField(CREATED, attrs.creationTime().toInstant());
-      }
-
-      if (params.shouldGetFileContent()) {
-        doc.setField(CONTENT, Files.readAllBytes(path));
-      }
-    } catch (Exception e) {
-      throw new ConnectorException("Error occurred getting/setting file attributes to document: " + path, e);
-    }
-    return doc;
-  }
-
-  @Override
-  protected boolean validFile(FileReference fileRef) {
-    // Currently a no-op. Files.walk won't return any directories
-    return true;
-  }
-
-  @Override
-  protected String getFullPath(FileReference fileRef, TraversalParams params) {
-    return fileRef.getPath().toString();
-  }
-
-  private Document pathToDoc(Path path, InputStream in, String decompressedFullPathStr, TraversalParams params)
-      throws ConnectorException {
-    String docId = DigestUtils.md5Hex(decompressedFullPathStr);
-    Document doc = Document.create(createDocId(docId, params));
-
-    try {
-      // get file attributes
-      BasicFileAttributes attrs = Files.readAttributes(path, BasicFileAttributes.class);
-
-      // setting fields on document
-      // remove Extension to show that we have decompressed the file and obtained its information
-      doc.setField(FILE_PATH, decompressedFullPathStr);
-      doc.setField(MODIFIED, attrs.lastModifiedTime().toInstant());
-      doc.setField(CREATED, attrs.creationTime().toInstant());
-      // unable to get decompressed file size
-      if (params.shouldGetFileContent()) {
-        doc.setField(CONTENT, in.readAllBytes());
-      }
-    } catch (Exception e) {
-      throw new ConnectorException("Error occurred getting/setting file attributes to document: " + path, e);
-    }
-    return doc;
   }
 
   public class LocalFileVisitor implements FileVisitor<Path> {
@@ -192,7 +180,7 @@ public class LocalStorageClient extends BaseStorageClient {
       Path fullPath = file.toAbsolutePath().normalize();
       String fullPathStr = fullPath.toString();
       String fileExtension = FilenameUtils.getExtension(fullPathStr);
-      tryProcessAndPublishFile(publisher, fullPathStr, fileExtension, new FileReference(fullPath, attrs.lastModifiedTime().toInstant()), params);
+      tryProcessAndPublishFile(publisher, fullPathStr, fileExtension, new LocalFileReference(fullPath, attrs.lastModifiedTime().toInstant()), params);
       return FileVisitResult.CONTINUE;
     }
